refactor: test are passing now
