+++ package.json
@@ -4,19 +4,17 @@
   "author": "jd-apprentice",
   "license": "MIT",
   "main": "dist/app/index.js",
+  "type": "module",
   "scripts": {
-    "prebuild": "rm -rf dist",
-    "build": "tsc",
-    "build:prod": "npm run prebuild && npm install && npm run build",
-    "dev": "NODE_ENV=development nodemon src/app/index.ts",
-    "start:dev": "NODE_ENV=development node .",
-    "start:prod": "NODE_ENV=production node .",
+    "build": "bun build src/app/index.ts --target=node --outfile dist/app/index.js",
+    "dev": "NODE_ENV=development bun --watch src/app/index.ts",
+    "start": "NODE_ENV=production node .",
     "lint": "eslint ./src/**/*.ts",
     "lint:fix": "eslint ./src/**/*.ts --fix",
     "format": "prettier --check ./src/**/*.ts",
     "format:fix": "prettier --write ./src/**/*.ts",
     "prepare": "husky",
-    "test": "jest"
+    "test": "bun test --coverage"
   },
   "dependencies": {
     "@hapi/boom": "^10.0.1",
@@ -34,32 +32,23 @@
     "rollbar": "^2.26.4"
   },
   "devDependencies": {
-    "@jest/globals": "^29.7.0",
-    "@nightwatch/apitesting": "^3.1.4",
     "@types/bcrypt": "^5.0.2",
+    "@types/bun": "^1.1.8",
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
-    "@types/jest": "^29.5.12",
     "@types/jsonwebtoken": "^8.5.9",
     "@types/morgan": "^1.9.9",
     "@types/multer": "^1.4.12",
-    "@types/node": "^17.0.45",
     "@types/supertest": "^6.0.2",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
+    "aws-sdk": "^2.1691.0",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-security": "^3.0.1",
     "eslint-plugin-sonarjs": "^2.0.2",
     "husky": "^9.1.5",
-    "jest": "^29.7.0",
-    "jest-html-reporters": "^3.1.7",
-    "jest-junit": "^16.0.0",
-    "nodemon": "^3.1.4",
+    "mock-aws-s3": "^4.0.2",
+    "nock": "^13.5.5",
     "prettier": "^3.3.3",
-    "supertest": "^6.3.4",
-    "ts-jest": "^29.2.5",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.5.4"
+    "supertest": "^6.3.4"
   }
 }
