+++ src/app/config/config.ts
@@ -1,7 +1,3 @@
-// External Modules
-import dotenv from "dotenv";
-dotenv.config();
-
 const {
   PORT,
   DB_HOST,
