+++ src/app/db/index.ts
@@ -2,8 +2,12 @@ import mongoose from "mongoose";
 import { rollbar } from "../config/rollbar";
 import Config from "../../app/config/config";
 
-const loadDatabase = async (uri: string): Promise<void> => {
+const loadDatabase = async (uri: string | undefined): Promise<void> => {
     try {
+        if (!uri) {
+            rollbar.error("URI not found");
+            return;
+        }
         const connection = await mongoose.connect(uri);
         if (Config.app.environment != "TEST") {
             rollbar.info(`Connected to ${connection.connection.name} database.`);
