+++ src/app/index.ts
@@ -2,7 +2,7 @@ import { app } from './main';
 import { loadDatabase } from './db';
 import Config from './config/config';
 
-app.listen(Config.app.port || 3000, async () => {
+app.listen(Config.app.port || 4000, async () => {
   await loadDatabase(Config.db.uri);
   console.log(`
         ██╗    ██╗ █████╗ ██╗███████╗██╗   ██╗██╗      █████╗ ███╗   ██╗██████╗      █████╗ ██████╗ ██╗
