--- tsconfig.build.json
@@ -1,9 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "exclude": [
-        "node_modules",
-        "__tests__",
-        "dist",
-        "**/*spec.ts"
-    ]
-}
